Functioning binary counter
The code can count in binary and decimal successfully. However, it will experience noise and change number randomly.
This feature is currently being tested in a moving car and so this may not occur in better conditions. But, one way to
improve this counter is by adding a filter that decreases the noise