Refactored and added more unit tests
This is being more thorough with the number of fingers being counted and the binary values returned.
More tests will be added to check the other fingers pointing down.

Finally more tests will be added that have the fingers pointing down. These will fail as the logic implemented currently
is incorrect. These tests should pass once the logic as been worked on