Created new function to calculate finger position
This function finds the fingers point relative to the wrist. This way it the finger tip can be compared
to the joints point, by getting their absolute value between the difference in their x/y position
relative to the wrist. This if the x/y value of the finger tip is greater than the finger joint the finger
can be said as being up