5 - Fixing issue with counting thumb in a diagonal position
To check the thumb is up, the function now checks that the thumb tip x/y value is larger than the x/y value of the index finger MCP joint.
The x/y depends on the angle of the hand. When the hand is upright the x-position is what is checked. Then when the hand is sideways the y-position
is checked. These values are relative to the pinky MCP joint. This is different to the other fingers as they are compared relative to the wrist.

Following this change the unit tests were altered to match the new changes to the function