Added unit test test cases for when the hand is diagonal
Two new test cases have been created for when the hand is diagonal. In the first case all fingers are up
including the thumb so all 0th index values within the returned list from the function should be greater
than the 1st index value

In the second test the thumb is down in the diagonal position. So the returned thumb list should have the 1st index
greater than the 0th index value