10: Preliminary fix idea
This is just a prototype idea to fix the binary counting issue. There is still more development that needs to occur.
As at the moment it will draw hands one colour, when two are in front of the camera. Additionally, still need to order the hands from left to right