4-set-up-serial-connection-to-arduino-device
Added basic functionality between program and arduino. It will display the counting type and number in one line as it is passed as one message.
I want to edit this as part of another issue. I want the top line to display whether it is binary or decimal. Then in the second line the value
that has been counted.