6-remove-binary-counter-when-hand-is-sideways-or-upside-down
The counting will adjust depending if the user is counting in binary or decimal. And it will ignore any hands that are sideways in binary.
So, if one hand is up right and the other is sideways it will ignore the sideways one. So, it will adjust to this and count one hand on binary.

I had to update the unit tests to account for this change as well. And they are still functioning as expected.

There was also an issue with the serial port. Where if there was no device attached to the specified port it would fail. So, now there is a check to make sure
there can be a connection made before attempting to open the port.